[refactor/#22] Refactor entities to extend BaseEntity
