+++ src/main/java/com/api/jaebichuri/screening/entity/AuctionScreening.java
@@ -1,14 +1,15 @@
-package com.api.jaebichuri.auction.entity;
+package com.api.jaebichuri.screening.entity;
 
 import com.api.jaebichuri.auction.enums.AuctionCategory;
-import com.api.jaebichuri.auction.enums.AuctionScreeningStatus;
+import com.api.jaebichuri.global.entity.BaseEntity;
 import com.api.jaebichuri.member.entity.Member;
+import com.api.jaebichuri.screening.enums.AuctionScreeningStatus;
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Getter
@@ -16,7 +17,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EntityListeners(AuditingEntityListener.class)
-public class AuctionScreening {
+public class AuctionScreening extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,10 +32,6 @@ public class AuctionScreening {
     @Column(nullable = false)
     private Long desiredStartPrice;
 
-    @CreatedDate
-    @Column(nullable = false)
-    private LocalDateTime createdAt;
-
     @Enumerated(EnumType.STRING)
     private AuctionScreeningStatus screeningStatus;
 
@@ -45,7 +42,16 @@ public class AuctionScreening {
     @JoinColumn(name = "member_id", nullable = false)
     private Member seller;
 
+    @OneToMany(mappedBy = "auctionScreening", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<AuctionScreeningImage> images = new ArrayList<>();
+
     public void setSeller(Member seller) {
         this.seller = seller;
     }
+
+    public void addImage(AuctionScreeningImage image) {
+        images.add(image);
+        image.setAuctionScreening(this);
+    }
+
 }
